feat: create useConfetti hook
